backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
pymongo==4.5.0
motor==3.1.1

# Authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI/ML
huggingface-hub==0.19.4
faiss-cpu==1.7.4
numpy>=1.21.0

# Utilities
pydantic==2.5.2
cachetools==5.3.2
python-jose==3.3.0
